src/apiappkv/requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
azure-core==1.32.0
azure-identity==1.21.0
azure-keyvault-secrets==4.9.0
psycopg2-binary==2.9.10